move Profile continent components into their own Profile folder; update profile photo
