+++ src/components/app/App.jsx
@@ -5,7 +5,7 @@ import {
   Switch
 } from 'react-router-dom';
 import Ocean from '../ocean/Ocean';
-import ProfileDetails from '../continents/ProfileDetails';
+import ProfileDetails from '../continents/Profile/ProfileDetails';
 import ProjectsDetails from '../continents/ProjectsDetails';
 import BeHumanDetails from '../projects/BeHuman/BeHumanDetails';
 import ChooseGifDetails from '../projects/ChooseGif/ChooseGifDetails';
